+++ CHANGELOG.md
@@ -16,6 +16,8 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - `cynic-introspection` can now output SDL
 - Added `CapabilityDetectionQuery` to `cynic-introspection`, which can detect
   which version of the GraphQL specification a server supports.
+- `QueryFragment` now allows users to specify features for parts of the query,
+  allowing the same query to be used on servers with differing capabilities.
 
 ## v3.1.1 - 2023-06-22